Refine benchmark tests
